fix(review): catch bootstrap failures behind the JSON contract, and cap positional arguments per command
